Rename RadioGroup to RadioButtonGroup
