+++ src/components/Radio.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from 'react';
 import styled from 'styled-components';
 import Box from './Box';
-import { RadioChecked, RadioEmpty } from './icons';
+import { RadioChecked, RadioEmpty } from '../icons';
 
 interface RadioProps {
   id: string;
